backend/requirements.txt
# ASGI Runtime — Container Apps
fastapi
uvicorn[standard]
python-multipart

# Azure Runtime
azure-identity
azure-keyvault-secrets

# Azure AI
openai
azure-ai-projects>=2.0.0b3
azure-ai-inference

# Microsoft Agent Framework SDK (sequential workflow)
# agent-framework --pre
agent-framework-azure-ai==1.0.0rc3
agent-framework-core==1.0.0rc3

# Azure Data Services
azure-cosmos
azure-storage-blob
azure-servicebus

# Azure AI Services
azure-ai-contentsafety
azure-search-documents
azure-cognitiveservices-speech

# Azure Voice Live SDK (real-time voice with Foundry Agent Service)
azure-ai-voicelive>=1.2.0b4
# pyaudio — only needed for local dev microphone testing, not in container

# Azure Web PubSub (real-time transport for SWA-compatible voice sessions)
azure-messaging-webpubsubservice

# Azure Document Intelligence + Computer Vision
azure-ai-documentintelligence
azure-ai-vision-imageanalysis

# Azure Communication Services (email + SMS reminders)
azure-communication-email
azure-communication-sms

# HTTP client for Video Indexer REST API
requests

# OpenTelemetry — App Insights SDK instrumentation
azure-monitor-opentelemetry
opentelemetry-instrumentation-fastapi

# Text analysis
textstat

# PDF fallback extraction (when Document Intelligence is unavailable)
PyPDF2

# Env loading (used by create_speech_agent.py)
python-dotenv

# Auth
PyJWT
cryptography

# PDF parsing
pypdf2
